requirements.txt
tensorflow==2.15.0
imutils==0.5.4
opencv-python==4.9.0.80
numpy==1.26.4
PyYAML>=6.0
